feat(ProductFilter): create search of products and select products by `userId`
Description:
- 4. Once this grid has been created, we must include a search engine at the top that allows us to filter these cards by searching for matches in title and/or description.

Steps:
- Create the `ProductFilter.vue` file as a new component.
- Import and use `ProductFilter` in the `ProductList` component.
- To use `FontAwesome` in vue, install and import npm packages, and create the configuration in the `main.js` file .
Create SCSS styles and use `fa-magnifying-glass` and `fa-chevron-down` icons for the search and select fields.
- Create `getUserIdList` as a computed property, to get a list of all possible user IDs.
And so, create a loop to print the `option` tags of the select field.
- Changes the import and use of the `ProductFilter` component, in the `ProductList` component to the `Product` view.
Now use `emit` vue to pass information from child components to parent components and `props` for parent to child.
Creates an `emitProductList` method to send the list of products from the `ProductList` component to the `Product` view and from the `Product` view to the `ProductFilter` component.
When the `ProductList` component is mounted and gets the data, the data from `ProductList` is sent to the `Product` view.
The `Product` view executes the `setProductList` method to store the data as `productList` and with the `getProductList` computed property sends the data back to the `ProductFilter` component.
- It creates a search engine in the `ProductFilter` component and the search results are sent to the `ProductList` component:
1. In the `ProductFilter` component, when typing in the input, the entered text is updated in the `productFilter.title` data. Then, the `emitProductFilterTitle` method sends the updated text to the `Products` view.
2. In the `Products` view, the `setProductFilterTitle` event saves the received text in the `productFilter.title` data and sends it to the `ProductList` component through the `product-filter-title` prop.
3. In the `ProductList` component, the `getProductListFilterByTitle` computed property uses the text stored in the `productFilterTitle` prop to filter the product list based on the text entered in the search engine.
- As with the search engine, create a selector to filter products by their `userId`.
- The `getProductList` computed property is still used to paint the data with the `v-for`, by default it always returns the `getProductList` computed property.
In case something has been searched by title, the `getProductListFilterByTitle` computed property is returned.
In case a user id has been selected, the computed property `getProductListFilterByUserId` is returned.
But, it is also possible to filter by both, by title and by user id.