fix(router): gets error when the path not exist
Error:
- Cannot read properties of undefined (reading 'id')

Solution:
- Define the `router` in `nuxt.config.js`.
- If the `currentLink` is `undefined`, it defaults to the `id` of the first object in the `linkList` array.
- Redirect to `/` if the path is `undefined`.
- To ensure that `this.$route` is defined before using `push`, it is possible that the Vue Router instance is not yet available at the time `setLinkIDDefaultActive` is called.
It could be to wait until the Vue Router instance is available before calling `setLinkIDDefaultActive`. This can be achieved by using the `created()` hook instead of the `mounted()` hook, since `created()` is called before the component has been mounted in the DOM and therefore before `this.$route` needs to be used. The `this.$nextTick()` hook waits for the DOM to update before calling `setLinkIDDefaultActive`.