feat(Login): create the `LoginForm` component and add it to the `Login` view
Steps to follow:
- Create the `LoginForm` component:
The `LoginForm` component is now included in the `Login` view.
It consists of email and password input fields.
When the form is filled in and the button is clicked, the system validates the credentials.
If the credentials are correct, the user is redirected to the `Account` page.
However, if the credentials are incorrect, an error message is displayed to the user.
- Update the `LoginView` view:
Import and use the `LoginForm` component in the `Login` view.