feat(SalesSection): creates the new section like the others.
Steps:
- Duplicate the `SectionOffers.vue` file and rename it to `SectionSales.vue`.
- Add the `SectionSales` component to the `ViewAccount` view.
- In the `SectionSales.vue` file:
Changes the name of the `SectionOffers` component to `SectionSales`.
Changes the title of the `Offers` data to `Points of sales data`.
Change the name of data `offersData` to `salesData`.
Change the name of the css class `offers` to `sales`.
Change a part of the url of `offers` to `sales_points`, used in the `fetchData` method.
Also update the error and success messages.