fix(SectionBusiness, SectionOffers, SectionSales): get console errors
Errors:
- [Vue warn]: Error in render: "TypeError: Failed to construct 'URL': Invalid URL"
- SectionBusiness.vue:306 Invalid URL: TypeError: Failed to construct 'URL': Invalid URL

Solutions:
- The error message suggests that there is an issue with the URL passed to the `URL` constructor. It is likely that the `url` parameter being passed to the `getPageNumberCurrent` function is not a valid URL.
To resolve this issue, you can check the value of the `url` parameter before using it to construct a `URL` object. So if an invalid URL is provided, an error message will be logged to the console, and the function will return an empty string. This will help you identify and handle any invalid URLs causing the error.
- The error message "Invalid URL: TypeError: Failed to construct 'URL': Invalid URL" is likely caused by passing an invalid URL to the `new URL(url)` constructor in the `getPageNumberCurrent` method.
The issue may occur if the `url` parameter passed to the `getPageNumberCurrent` method is not a valid URL. This can happen if `getBusinessData.next` is not a valid URL or is empty.
To avoid the error, you can add a check to ensure that `url` is a valid string before attempting to construct a URL object. Added a check `if (url)` to verify that `url` is not empty or undefined before attempting to construct a URL object. If `url` is empty, it will return an empty string as before.
- Remove duplicated class `business` form the navigation.
- Rename `businessesData` to `businessData`.