fix(UISection): current page is 1 when on the last page
Error:
- When the last page is reached by pressing the `Next` button, the path obtained from the data does not exist, i.e. it is `null`.
In that case the `getResultPage` method is incorrectly configured to always return 1 as the current page.

Solution:
- Convert the `resultPage` data into an object that stores the total number of pages, the previous page number, the next page number, and the current page number.
This will allow easier access to these values when new data is obtained.
- Call the `resultPage` method inside the `setData` method instead of directly in the HTML.
The `setData` method is called every time new data is requested using the `fetchData` method, ensuring that the current page is updated accordingly.
- In the HTML, use the `resultPage` data to display the current page and enable/disable the previous and next buttons based on the availability of the previous and next pages.
- Update the `getResultIndex` method, since `resultPage` is now an object and before it was directly the page number.
- Update the `getResultPage` method to handle different scenarios:
If no path is available, the value will be `null`.
If the path is available but there is no page number, the value will be the first and last page, i.e. 1 for the previous page and the total number of pages for the next page.
If the path is available and there is a page number, the current page will be the result of adding 1 to the previous page.