fix(UISectionFIlter): incorrect update filter by sector when change page
Error:
- From the `UISection` component, when the `UISectionFilter` `sectorSelected` prop value is updated, the `sectorSelectedSaved` data is not updated.

Solution:
- To make sure that the `sectorSelectedSaved` property is updated when the `sectorSelected` property changes, you can use a watch function in the `UISectionFilters` component.
A `watch` block is defined in the `UISectionFilters` component that looks for changes in the `sectorSelected` property. When a change in `sectorSelected` is detected, the `watch` function is executed and updates the value of `sectorSelectedSaved` with the new value of `sectorSelected`.
This way, whenever the `sectorSelected` property changes outside the `UISectionFilters` component, the `sectorSelectedSaved` property will be automatically updated to reflect the new value.