improve(UISectionProperty): load a default file when no icon or image is found
Description:
- The `UISectionProperty` component used in `summary`, uses the `propertyKey` property to send different data than expected. The `propertyKey` property was created to display the `key` property, but in this case it is desired to display a different `key`.
To accommodate this possibility, the `propertyIndex` property has been introduced in the `UISectionProperty` component so that you always have access to the `key` even if you decide to pass another text for the `strong` tag.
This allows the component to check, in case the `value` is empty, the text of the `key`, and thus handle cases where an image or icon is missing by displaying a default file when necessary.

Steps:
- Create default files for the icons and images:
Save the default icon and image files in the `assets` folder.
Import these files in the `UISectionProperty` component for use.
In the `UISectionProperty` component, create the `getIconDefault` and `getImageDefault` computed properties to get the default files and use them in the `img` html tag.
- In the `UISectionProperty` component:
Rename `propertyKey` prop to `propertyName`.
Create `propertyIndex` prop and set its default value to `propertyName`.
In the condition where the `value` of the property is empty, check if its `key` corresponds to an icon or image.
- In the `summary` section inside the `UISectionArticle` component:
Add to the `UISectionProperty` component the `propertyIndex` prop as `icon`, since in this particular case we know what its `key` is.