refactor(UISectionArticle): organizing nested `v-for` parameters
Description:
- This commit refactors the `UISectionArticle` component by renaming the parameters of each nested `v-for` loop.
The objective is to create a clear hierarchy by facilitating the understanding of nested loops, to improve the organization, readability and maintainability of the code.

Steps:
- For the first iteration level:
Rename `propertyValue` to `propertyValueLevel1`.
Rename `propertyName` to `propertyNameLevel1`.
- For the second iteration level:
Rename `dataValue` to `propertyValueLevel2`.
Rename `dataIndex` to `propertyNameLevel2`.
- For the third iteration level:
Rename `itemObj` to `propertyValueLevel3`.
Rename `indexObj` to `propertyNameLevel3`.
- For the fourth iteration level
Rename `item` to `propertyValueLevel4`.
Rename `index` to `propertyNameLevel4`.