fix(data): missing data when api is not working
Error:
- When the API is down because of not being able to access the data, no further work can be done.

Reason:
- For that reason temporary files of that data were created in the `assets/data` folder. So when the request fails, in each `UISection` with the `dataFileTest` prop, the json of the temporary file is obtained.
But not being complete some parts of the code break, like pagination, since with each click fetch requests are made to new data that do not exist.

Solution:
- At this moment that the API is working I took the opportunity to download the data.
So I created the function `downloadData`, to avoid having to download them manually because there are many pages.
That function is executed during the request of the fetch, after obtaining them it uses the `file-server` package to create a json file with the name of the page number.
In a web browser environment, JavaScript does not have direct access to the user's file system for security reasons. This means that you cannot specify a specific location to save a file to the user's hard drive from JavaScript in a browser.
When you use the `saveAs` function of `FileSaver.js`, the file is automatically saved to the user's default download folder as configured in their browser.