fix(UISectionArticleContent): the `is_favorite` property is not updated
Reason:
- When variables are declared in the `data` block, they are initialized when the component is created and the initial value of `resultValue` is assigned to `resultValueSaved`. At that time, the value of `resultValue` may not yet be fully loaded or may have subsequently changed.

Solution:
- A computed property called `sortedResultValue` has been added that uses the `sortProperties` global method to sort `resultValueSaved` properties.
When using a computed property, such as `sortedResultValue`, it is evaluated dynamically whenever one of the dependencies changes. In this case, the dependency is `resultValueSaved`, and when `resultValueSaved` changes due to `watch`, the computed property is updated and the properties are sorted according to the `sortProperties` function.
This way, you make sure that `resultValueSaved` has the updated value and then sort it in the computed property `sortedResultValue`. By using `sortedResultValue` in the `v-for` loop of the template, you are iterating over the already sorted properties.